Enable toadserver to run as a service
Add a Dockerfile and start.sh script to enable the toadserver to
be built as a docker image suitable for use as an Eris service.
The Dockerfile and toadserver.toml file make some assumptions
about where the docker image is hosted and the github path.

Known issues:
- Environment variables must be provided for toadserver to
work correctly; these vary by environment and must be set
in the start script.

- Service does not stop from `eris stop toadserver`.